+++ example/test_driver/steps/tap_button_n_times_step.dart
@@ -8,8 +8,8 @@ class TapButtonNTimesStep extends When2WithWorld<String, int, FlutterWorld> {
   @override
   Future<void> executeStep(String input1, int input2) async {
     final locator = find.byValueKey(input1);
-    for (var i = 0; i < 10; i += 1) {
-      await world.driver.tap(locator, timeout: timeout);
+    for (var i = 0; i < input2; i += 1) {
+      await FlutterDriverUtils.tap(world.driver, locator, timeout: timeout);
     }
   }